phase2_hack2/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
psycopg[binary]==3.1.18
asyncpg==0.29.0
aiosqlite==0.22.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
alembic==1.13.1
